requirements.txt
django
djangorestframework
django-cors-headers
google-generativeai
paddleocr
paddlepaddle
opencv-python
pillow